+++ src/tsx/components/main/PressContainer.tsx
@@ -3,8 +3,8 @@ import gridIcon from "../../../img/gridIcon.svg";
 import listIcon from "../../../img/listIcon.svg";
 import styled from "styled-components";
 import { MENU_STATES, News, SubscribeAction, VIEW_STATES } from "../../constants";
-import GridView from "./GridView";
-import ListView from "./ListView";
+import GridView from "./grid/GridView";
+import ListView from "./list/ListView";
 import { NewsContext } from "../provider/NewsProvider";
 import { deleteSubscription, fetchSubscription, postSubscription } from "../../api/NewsAPI";
 import { SubscribeContext } from "../provider/SubscribeProvider";
